+++ src/pages/api/v1/users/[key].ts
@@ -2,9 +2,9 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import z from 'zod';
 
 import { handleRouteErrors, requirePutMethod } from '@/server/helpers/server';
+import { upsertUsersFromApi } from '@/server/services/airtable';
 import { apiUser } from '@/services/api/authentication';
 import { withCors } from '@/services/api/cors';
-import { upsertUsersFromApi } from '@/services/users';
 
 const ApiNetworkValidation = z.object({
   id_sncu: z.string(),
